Migrate some 2.0 types out of api
